multi-threading, aside from responses
The last real feature I am adding before finishing this version of the project is multi-threading. This completes multi-threading, essentially -- but there is a problem. We don't seem to be able to respond, similarly to the case with `Toolips`.... So I have begun the process of making a `UDPIOConnection`, an equivalent to the `IOConnection` used in `Toolips` -- as this is the same exact problem that app was faced with c;