finally got multi-threading completely working....
Unfortunately, like `Toolips` I ended up having to create an `IOConnection` and then respond with its Stream. At least the servers on this front are a bit more open-ended. 
I want to curse for this one -- this was a *pain in the ass*, lol. I have no idea why my current solution worked whereas the numerous others did not. I kept continuously getting looped with a single client -- likely something was *actually confusing julia/sockets* with the weird scope of the loop and the `new_job` being made there. All sorts of weirdness on top of the whole process being genuinely outrageous. I don't think there is a `found` exception for the `Bool` returns in the threaded `job` yet.